ci: stop if running

+++ vhdbuilder/packer/sysprep.ps1
@@ -1,6 +1,11 @@
 # Stop and remove Azure Agents to enable use in Azure Stack
 # If deploying an Azure VM the agents will be re-added to the VMs at deployment time
-Stop-Service WindowsAzureGuestAgent
+$ServiceName = 'WindowsAzureGuestAgent'
+$arrService = Get-Service -Name $ServiceName
+if ($arrService.Status -eq 'Running'){
+    echo "$ServiceName is still running, trying to stop now."
+    Stop-Service WindowsAzureGuestAgent
+}
 Stop-Service RdAgent
 & sc.exe delete WindowsAzureGuestAgent
 & sc.exe delete RdAgent
